refactor: separate comparators and EvseSecurity mock into reusable headers
